Add files via upload
In the model, mobilenet v2 pretrained model was used. An image of the eye retina is taken as an input. The model predicts whether the retina has symptoms of diabetic retinopathy and if the retina has the symptoms, the model predicts the level of diabetes. The train accuracy is 72% and the test accuracy is 54%.